refactor: 2차 리팩토링
